remove abstract class for e2e tests
