+++ src/test/java/com/zufar/icedlatte/product/endpoint/ProductsEndpointTest.java
@@ -1,18 +1,25 @@
 package com.zufar.icedlatte.product.endpoint;
 
 import com.zufar.icedlatte.product.util.PaginationAndSortingAttribute;
-import com.zufar.icedlatte.test.config.AbstractE2ETest;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.data.domain.Sort;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
+import io.restassured.specification.RequestSpecification;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.data.domain.Sort;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static com.zufar.icedlatte.test.config.RestAssertion.assertRestApiBadRequestResponse;
 import static com.zufar.icedlatte.test.config.RestAssertion.assertRestApiNotFoundResponse;
@@ -22,7 +29,23 @@ import static io.restassured.RestAssured.given;
 
 @Testcontainers
 @DisplayName("ProductsEndpoint Tests")
-class ProductsEndpointTest extends AbstractE2ETest {
+@ActiveProfiles("test")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class ProductsEndpointTest {
+
+    @Container
+    @ServiceConnection
+    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:13.11-bullseye");
+
+    @LocalServerPort
+    protected Integer port;
+
+    @DynamicPropertySource
+    static void dataSourceProperties(DynamicPropertyRegistry registry) {
+        registry.add("DATASOURCE_URL", postgres::getJdbcUrl);
+    }
+
+    protected static RequestSpecification specification;
 
     private static final String PRODUCT_SCHEMA_LOCATION = "product/model/schema/product-schema.json";
     private static final String PRODUCT_FAILED_SCHEMA_LOCATION = "product/model/schema/product-failed-schema.json";
